Revert "Merge branch 'dev_jj_3.2' into WoutBranch"
This reverts commit b586f1867f2019c91132a49dab1e9de83b979867, reversing
changes made to c1b9523dd42fc201ba9a7065f43e77321bbf6268.

+++ myWebApp/Models/WorkspaceModel.cs
@@ -20,9 +20,9 @@ namespace myWebApp.Models
         [BindProperty]
         public int SquareMeters { get; set; } = 1;
         [BindProperty]
-        public double Lengthws { get; set; } = 1;
+        public int Lengthws { get; set; } = 1;
         [BindProperty]
-        public double Widthws { get; set; } = 1;
+        public int Widthws { get; set; } = 1;
         [BindProperty]
         public int Id {get; set;}
     }
